Change *grouping* to *set* for disambiguation
I thought that it may be worth noting that I am reading the docs and when reading "A tensor consists of a set of primitive values shaped into an array", it was ambiguous whether or not the ordering in the array matters. I assume that - at least in some cases - the ordering matters, and that maybe the word "set" could be more clear as something akin to a "grouping".